backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AI Detection Dependencies
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
alembic==1.13.1

# File processing and storage
python-multipart==0.0.6
aiofiles==23.2.1
minio==7.2.0
PyMuPDF==1.26.3
python-docx==1.2.0
Pillow==10.3.0

# PDF and document processing - enhanced for evaluation
pdf2image==1.17.0
pytesseract==0.3.10
docx2txt==0.9
pdfplumber==0.11.7
reportlab==4.0.7

# AI and LLM integration - enhanced for evaluation
openai==1.98.0
httpx==0.25.2
tiktoken==0.5.2
requests==2.31.0

# Data processing and utilities
pandas==2.1.4
numpy==1.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Background tasks and caching
celery==5.3.4
redis==5.0.1

# Logging and monitoring
structlog==23.2.0
sentry-sdk==1.39.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

